+++ flutter_inappwebview_android/android/src/main/java/com/pichillilorenzo/flutter_inappwebview_android/webview/in_app_webview/InAppWebView.java
@@ -241,8 +241,22 @@ final public class InAppWebView extends InputAwareWebView implements InAppWebVie
     }
   }
 
+  @Override
+  public void setAlpha(float alpha) {
+    ViewParent parent = getParent();
+    if (parent instanceof PullToRefreshLayout) {
+      ((PullToRefreshLayout) parent).setAlpha(alpha);
+    } else {
+      super.setAlpha(alpha);
+    }
+  }
+
   @SuppressLint("RestrictedApi")
   public void prepare() {
+    if (customSettings.alpha != null) {
+      setAlpha(customSettings.alpha.floatValue());
+    }
+
     javaScriptBridgeEnabled = customSettings.javaScriptBridgeEnabled;
     if (customSettings.javaScriptBridgeOriginAllowList != null && customSettings.javaScriptBridgeOriginAllowList.isEmpty()) {
       // an empty list means that the JavaScript Bridge is not allowed for any origin.
@@ -328,7 +342,8 @@ final public class InAppWebView extends InputAwareWebView implements InAppWebVie
     settings.setLoadWithOverviewMode(customSettings.loadWithOverviewMode);
     settings.setUseWideViewPort(customSettings.useWideViewPort);
     settings.setSupportZoom(customSettings.supportZoom);
-    settings.setTextZoom(customSettings.textZoom);
+    if (customSettings.textZoom != null)
+      settings.setTextZoom(customSettings.textZoom);
 
     setVerticalScrollBarEnabled(!customSettings.disableVerticalScroll && customSettings.verticalScrollBarEnabled);
     setHorizontalScrollBarEnabled(!customSettings.disableHorizontalScroll && customSettings.horizontalScrollBarEnabled);
@@ -898,7 +913,7 @@ final public class InAppWebView extends InputAwareWebView implements InAppWebVie
     if (newSettingsMap.get("supportZoom") != null && customSettings.supportZoom != newCustomSettings.supportZoom)
       settings.setSupportZoom(newCustomSettings.supportZoom);
 
-    if (newSettingsMap.get("textZoom") != null && !customSettings.textZoom.equals(newCustomSettings.textZoom))
+    if (newSettingsMap.get("textZoom") != null && (customSettings.textZoom == null || !customSettings.textZoom.equals(newCustomSettings.textZoom)))
       settings.setTextZoom(newCustomSettings.textZoom);
 
     if (newSettingsMap.get("verticalScrollBarEnabled") != null && customSettings.verticalScrollBarEnabled != newCustomSettings.verticalScrollBarEnabled)
@@ -1517,6 +1532,10 @@ final public class InAppWebView extends InputAwareWebView implements InAppWebVie
 
   @Override
   public boolean onTouchEvent(MotionEvent ev) {
+    if (!customSettings.isUserInteractionEnabled) {
+      return true;
+    }
+
     lastTouch = new Point((int) ev.getX(), (int) ev.getY());
 
     ViewParent parent = getParent();